Remove nameof(argumentName) macro
Because it doesn't make sense to say "argumentName is null" or "argumentName is less than 0". The 'nameof' macro should be used when implementing these functions, not inside them.

This is technically a bug, since `GuardClause.IsEmptyGuid(new Guid(), "Hello World");` would throw an exception with the message "argumentName can not be a string with value of empty guid.". Notice "Hello World" doesn't appear in the message because the actual argument's value isn't used, but rather it's name.